+++ caikit/core/__init__.py
@@ -30,13 +30,12 @@ import warnings as _warnings
 _warnings.filterwarnings("ignore")
 
 # Local
-from . import data_model, module, module_config, resources, toolkit, workflows
 from .data_model import DataObjectBase, dataobject
 from .model_manager import *
 from .module import *
-from .module_backend_config import configure as backend_configure
+from caikit.core.module_backends.module_backend_config import configure as backend_configure
 from .module_backends import *
-from .module_config import ModuleConfig
+from caikit.core.modules.config import ModuleConfig
 from .resources.base import ResourceBase, resource
 from .task import TaskBase, task
 from .toolkit import *
