+++ tests/core/test_model_manager.py
@@ -22,9 +22,7 @@ import tempfile
 import uuid
 
 # Local
-from caikit.core.module_backend_config import configure, configured_load_backends
-from caikit.core.module_backends import LocalBackend
-from caikit.core.module_backends.base import SharedLoadBackendBase
+from caikit.core.module_backends.module_backend_config import configure, configured_load_backends
 
 # Unit Test Infrastructure
 from sample_lib.modules.sample_task import SampleModule
@@ -177,17 +175,14 @@ class TestModelManager(TestCaseBase):
     def test_import_block_registry(self):
         """Make sure that the BLOCK_REGISTRY can be imported from model_manager"""
         # pylint: disable = import-outside-toplevel,no-name-in-module,unused-import
-        from caikit.core.model_manager import BLOCK_REGISTRY  # isort: skip
 
     def test_import_workflow_registry(self):
         """Make sure that the WORKFLOW_REGISTRY can be imported from model_manager"""
         # pylint: disable = import-outside-toplevel,no-name-in-module,unused-import
-        from caikit.core.model_manager import WORKFLOW_REGISTRY  # isort: skip
 
     def test_import_resource_registry(self):
         """Make sure that the RESOURCE_REGISTRY can be imported from model_manager"""
         # pylint: disable = import-outside-toplevel,no-name-in-module,unused-import
-        from caikit.core.model_manager import RESOURCE_REGISTRY  # isort: skip
 
 
 # Pytest tests #########################################################
